Subscribe to boot environment property changes to keep UI in sync.
This commit adds a subscription for all PropertiesChanged signals in the
boot environment namespace (making use of a D-Bus match rule) so that
the UI stays in sync.

At present we just reload everything when we detect changes (which
should be rare), but we could be more careful and add partial updates
and debouncing in the future.